requirements.txt
fastapi==0.115.7
pydantic==2.10.5
redis==5.2.1
SQLAlchemy==2.0.36
starlette==0.41.3
uvicorn==0.34.0
pymysql==1.1.1
itsdangerous==2.2.0
prometheus-fastapi-instrumentator==7.0.2
python-dotenv==1.0.1